refactor: move ip feature to net package
