+++ internal/service/net/ip_addr.go
@@ -1,4 +1,4 @@
-package web
+package net
 
 import (
 	"errors"
